🧹 Refactor dataset handling to introduce CollateMode enum, enhancing type safety and clarity in collate function usage across OTXDataset and related classes.

+++ src/otx/core/data/dataset/detection.py
@@ -31,7 +31,7 @@ class OTXDetectionDataset(OTXDataset):
             else np.zeros((0, 4), dtype=np.float32)
         )
 
-        labels = np.array([ann.label for ann in bbox_anns], dtype=np.int32)
+        labels = np.array([ann.label for ann in bbox_anns], dtype=np.int64)
 
         img_info = ImageInfo(
             img_idx=index,
